+++ include/01_forward_config.cuh
@@ -49,5 +49,5 @@ struct KernelConfig {
         alignas(16) float  row_sum[BLOCK_M];
     };
 
-    static constexpr size_t TOTAL_SMEM = ((sizeof(SmemLayout) + 127) & ~size_t(127));
+    static constexpr int TOTAL_SMEM = static_cast<int>(((sizeof(SmemLayout) + 127) & ~size_t(127)));
 };
